requirements.txt
asgiref
Django
djangorestframework
django-cors-headers
django-filter
django-debug-toolbar
djangorestframework-simplejwt
PyJWT
pandas
python-dotenv